get_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
weaviate-client
openai
faiss-cpu
numpy
sentence-transformers